+++ CHANGELOG.md
@@ -5,12 +5,20 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][Unreleased]
 
+
+
+
+* * *
+
+## [0.29.0][0.29.0] - 2017-10-22
+
 Added:
 
 1. Add Bot API v3.2 methods:
   * (#429) *TelegramBot#getStickerSet* (by @CapacitorSet, @LibertyLocked)
   * (#430) *TelegramBot#uploadStickerFile* (by @CapacitorSet)
   * *TelegramBot#createNewStickerSet*, *TelegramBot#addStickerToSet*, *TelegramBot#setStickerPositionInSet*, *TelegramBot#deleteStickerFromSet* (by @GochoMugo)
+1. Supports API v3.3
 
 Deprecated:
 
@@ -160,4 +168,5 @@ Fixed:
 [0.27.0]:https://github.com/yagop/node-telegram-bot-api/releases/tag/v0.27.0
 [0.27.1]:https://github.com/yagop/node-telegram-bot-api/releases/tag/v0.27.1
 [0.28.0]:https://github.com/yagop/node-telegram-bot-api/releases/tag/v0.28.0
-[Unreleased]:https://github.com/yagop/node-telegram-bot-api/compare/v0.27.1...master
+[0.29.0]:https://github.com/yagop/node-telegram-bot-api/releases/tag/v0.29.0
+[Unreleased]:https://github.com/yagop/node-telegram-bot-api/compare/v0.29.0...master